Rewrote AbstractFileStorage methods, added vararges in model's classes methods.

+++ src/ru/javawebinar/basejava/storage/AbstractFileStorage.java
@@ -14,11 +14,11 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
     protected File directory;
 
     protected AbstractFileStorage(File directory) {
-        Objects.requireNonNull(directory, "directory must not be null");
+        Objects.requireNonNull(directory, " directory must not be null");
         if (!directory.isDirectory()) {
-            throw new IllegalArgumentException(directory.getAbsolutePath() + "is not directory");
+            throw new IllegalArgumentException(directory.getAbsolutePath() + " is not directory");
         } else if (!directory.canRead() || !directory.canWrite()) {
-            throw new IllegalArgumentException(directory.getAbsolutePath() + "is not readable|writebal");
+            throw new IllegalArgumentException(directory.getAbsolutePath() + " is not readable|writebal");
         }
         this.directory = directory;
     }
@@ -30,7 +30,11 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
 
     @Override
     protected Resume getItem(File file) {
-        return read(file);
+        try {
+            return read(file);
+        } catch (IOException e) {
+            throw new StorageException("Can't read " + file.getAbsolutePath(), file.getName(), e);
+        }
     }
 
     @Override
@@ -39,13 +43,17 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
             file.createNewFile();
             write(file, resume);
         } catch (IOException e) {
-            throw new StorageException("IO error", file.getName(), e);
+            throw new StorageException("Can't create " + file.getAbsolutePath(), file.getName(), e);
         }
     }
 
     @Override
     protected void deleteItem(File file) {
-        file.delete();
+        if (!file.delete()) {
+            throw new StorageException("File delete error ", file.getName());
+        } else {
+            file.delete();
+        }
     }
 
     @Override
@@ -53,7 +61,7 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
         try {
             write(file, resume);
         } catch (IOException e) {
-            throw new StorageException("IO error", file.getName(), e);
+            throw new StorageException("Can't update " + file.getAbsolutePath(), file.getName(), e);
         }
     }
 
@@ -61,7 +69,7 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
     protected List<Resume> getList() {
         List<Resume> resumes = new ArrayList<>();
         for (File file : getFiles()) {
-            resumes.add(read(file));
+            resumes.add(getItem(file));
         }
         return resumes;
     }
@@ -69,31 +77,35 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
     @Override
     public void clear() {
         for (File file : getFiles()) {
-            file.delete();
+            deleteItem(file);
         }
     }
 
     @Override
     public int size() {
-        return getFiles().size();
+        return Objects.requireNonNull(directory.listFiles()).length;
     }
 
     protected List<File> getFiles() {
-        List<File> getFiles = new ArrayList<>();
-        for (File file : Objects.requireNonNull(directory.listFiles())) {
-            if (file.isFile()) {
-                getFiles.add(file);
+        if (directory.listFiles() != null) {
+            List<File> getFiles = new ArrayList<>();
+            for (File file : Objects.requireNonNull(directory.listFiles())) {
+                if (file.isFile()) {
+                    getFiles.add(file);
+                }
             }
+            return getFiles;
+        } else {
+            throw new StorageException("Directory read error ", null);
         }
-        return getFiles;
     }
 
     @Override
     protected boolean isItemExist(File file) {
-        return file.isFile();
+        return file.exists();
     }
 
-    protected abstract Resume read(File file);
+    protected abstract Resume read(File file) throws IOException;
 
     protected abstract void write(File file, Resume resume) throws IOException;
 }
